Adding option for a second environment secret
Many APIs have an access key and a secret key, both  of which would be good to have as secrets and remove from disclosure in code. Adding the option for a second  environment secret in the parent workflow makes it so that developers can have more accessibility. I've named it key, just to be generic.